requirements.txt
fastapi
uvicorn
motor
passlib[bcrypt]
pyjwt
pydantic[email]
python-multipart
httpx
bcrypt==4.0.1
python-dotenv
asyncssh
fastapi-mail